+++ apps/client/src/app/(auth)/products/datamarketplace/dataset/page.tsx
@@ -32,7 +32,7 @@ export default function Page() {
     navigator.clipboard.writeText(datasetId ?? "")
     toast({
       title: "Notification",
-      description: <p className="text-neutral-600">{uiConstants.copiedToClipBoard}</p>
+      description: <p className="text-stone-600">{uiConstants.copiedToClipBoard}</p>
     })
   }
 
@@ -43,13 +43,13 @@ export default function Page() {
         <TableCell className="hidden md:table-cell">{ds?.rating}</TableCell>
         <TableCell>
           <Suspense condition={ds?.rating >= 4.5} fallback={null}>
-            <Badge variant="default" key={"gold"}><Medal className="scale-50" />Gold</Badge>
+            <Badge className="gold-badge" key={"gold"}><Medal className="scale-50" />Gold</Badge>
           </Suspense>
           <Suspense condition={ds?.rating >= 4.0 && ds?.rating < 4.5} fallback={null}>
-            <Badge variant="secondary" key={"silver"}><Medal className="scale-50" />Silver</Badge>
+            <Badge className="silver-badge" key={"silver"}><Medal className="scale-50" />Silver</Badge>
           </Suspense>
           <Suspense condition={ds?.rating < 4.0} fallback={null}>
-            <Badge variant="outline" key={"bronze"}><Medal className="scale-50" />Bronze</Badge >
+            <Badge className="bronze-badge" key={"bronze"}><Medal className="scale-50" />Bronze</Badge >
           </Suspense>
         </TableCell>
         <TableCell className="text-right hidden md:table-cell">
@@ -216,7 +216,7 @@ export default function Page() {
                   </div>
                 </CardContent>
                 <CardFooter className="flex flex-row items-center bg-muted/50 px-6 py-3">
-                  <Button variant="default" className="w-full" onClick={(): void => router.push("/apireference")}>Data API Reference<BookMarked className="scale-75" /></Button>
+                  <Button variant="default" className="w-full" onClick={(): void => router.push("/apireference?tab=datamarketplace")}>Data API Reference<BookMarked className="scale-75" /></Button>
                 </CardFooter>
               </Card>
             </div>