+++ apps/client/src/components/infopanel/index.tsx
@@ -21,7 +21,7 @@ export default function InfoPanel({ title, desc, value, masked, capitalize }: In
     navigator.clipboard.writeText(value)
     toast({
       title: "Notification",
-      description: <p className="text-neutral-600">{uiConstants.copiedToClipBoard}</p>
+      description: <p className="text-stone-600">{uiConstants.copiedToClipBoard}</p>
     })
   }