+++ apps/server/src/products/httpnosql/commands/impl/create-data.command.ts
@@ -2,6 +2,6 @@ export class CreateDataCommand {
   constructor(
     public readonly orgId: string,
     public readonly key: string,
-    public readonly value: string
+    public readonly value: Record<string, any> | Record<string, any>[] | string | string[]
   ) { }
 }